[Refactor] Move ChunkAccumulator and its test to Runtime
