+++ be/src/exec/pipeline/nljoin/nljoin_context.cpp
@@ -26,8 +26,8 @@
 #include "exec/pipeline/runtime_filter_hub.h"
 #include "exprs/expr.h"
 #include "fmt/format.h"
+#include "runtime/chunk_accumulator.h"
 #include "runtime/runtime_state.h"
-#include "storage/chunk_helper.h"
 
 namespace starrocks::pipeline {
