+++ be/src/exec/pipeline/nljoin/nljoin_context.h
@@ -29,8 +29,8 @@
 #include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/spill_process_channel.h"
 #include "exprs/expr_context.h"
+#include "runtime/chunk_accumulator.h"
 #include "runtime/runtime_state_fwd.h"
-#include "storage/chunk_helper.h"
 
 namespace starrocks {
 class RuntimeFilterBuildDescriptor;
