+++ be/src/exec/pipeline/scan/schema_chunk_source.h
@@ -19,8 +19,8 @@
 #include "exec/pipeline/scan/chunk_source.h"
 #include "exec/pipeline/scan/scan_operator.h"
 #include "exec/pipeline/scan/schema_scan_context.h"
+#include "runtime/chunk_accumulator.h"
 #include "runtime/runtime_state_fwd.h"
-#include "storage/chunk_helper.h"
 
 namespace starrocks {
 class SchemaScannerParam;
