+++ be/src/exprs/array_map_expr.cpp
@@ -33,7 +33,7 @@
 #include "exprs/expr_context.h"
 #include "exprs/function_helper.h"
 #include "exprs/lambda_function.h"
-#include "storage/chunk_helper.h"
+#include "runtime/chunk_accumulator.h"
 
 namespace starrocks {
 ArrayMapExpr::ArrayMapExpr(const TExprNode& node) : Expr(node, false) {}
