+++ be/src/exprs/array_sort_lambda_expr.cpp
@@ -35,7 +35,6 @@
 #include "exprs/expr_context.h"
 #include "exprs/function_helper.h"
 #include "exprs/lambda_function.h"
-#include "storage/chunk_helper.h"
 
 namespace starrocks {
 ArraySortLambdaExpr::ArraySortLambdaExpr(const TExprNode& node) : Expr(node, false) {}
