+++ be/src/storage/chunk_helper.cpp
@@ -26,7 +26,6 @@
 #include "column/chunk.h"
 #include "column/column_helper.h"
 #include "column/column_visitor_adapter.h"
-#include "column/json_column.h"
 #include "column/map_column.h"
 #include "column/runtime_type_traits.h"
 #include "column/schema.h"
@@ -35,7 +34,7 @@
 #include "column/vectorized_fwd.h"
 #include "exprs/expr_context.h"
 #include "gutil/strings/fastmem.h"
-#include "runtime/current_thread.h"
+#include "runtime/chunk_accumulator.h"
 #include "storage/tablet_schema.h"
 #include "types/olap_type_infra.h"
 #include "types/storage_type_traits.h"
@@ -161,105 +160,6 @@ void ChunkHelper::padding_char_columns(const std::vector<size_t>& char_column_in
         Column* column = chunk->get_column_raw_ptr_by_index(field_index);
         padding_char_column(tschema, *schema.field(field_index), column);
     }
-}
-
-ChunkAccumulator::ChunkAccumulator(size_t desired_size) : _desired_size(desired_size) {}
-
-void ChunkAccumulator::set_desired_size(size_t desired_size) {
-    _desired_size = desired_size;
-}
-
-void ChunkAccumulator::reset() {
-    _output.clear();
-    _tmp_chunk.reset();
-    _accumulate_count = 0;
-}
-
-namespace {
-bool check_json_schema_compatibility(const Chunk* one, const Chunk* two) {
-    if (one->num_columns() != two->num_columns()) {
-        return false;
-    }
-
-    for (size_t i = 0; i < one->num_columns(); i++) {
-        auto& c1 = one->get_column_by_index(i);
-        auto& c2 = two->get_column_by_index(i);
-        const auto* a1 = ColumnHelper::get_data_column(c1.get());
-        const auto* a2 = ColumnHelper::get_data_column(c2.get());
-
-        if (a1->is_json() && a2->is_json()) {
-            auto json1 = down_cast<const JsonColumn*>(a1);
-            if (!json1->is_equallity_schema(a2)) {
-                return false;
-            }
-        } else if (a1->is_json() || a2->is_json()) {
-            // never hit
-            DCHECK_EQ(a1->is_json(), a2->is_json());
-            return false;
-        }
-    }
-
-    return true;
-}
-} // namespace
-
-Status ChunkAccumulator::push(ChunkPtr&& chunk) {
-    size_t input_rows = chunk->num_rows();
-    // TODO: optimize for zero-copy scenario
-    // Cut the input chunk into pieces if larger than desired
-    for (size_t start = 0; start < input_rows;) {
-        size_t remain_rows = input_rows - start;
-        size_t need_rows = 0;
-        if (_tmp_chunk) {
-            need_rows = std::min(_desired_size - _tmp_chunk->num_rows(), remain_rows);
-            // Check JSON schema compatibility before appending
-            if (!check_json_schema_compatibility(_tmp_chunk.get(), chunk.get())) {
-                // Schema mismatch, output current chunk and create a new one
-                _output.emplace_back(std::move(_tmp_chunk));
-                _tmp_chunk = chunk->clone_empty(_desired_size);
-                TRY_CATCH_BAD_ALLOC(_tmp_chunk->append(*chunk, start, need_rows));
-            } else {
-                TRY_CATCH_BAD_ALLOC(_tmp_chunk->append(*chunk, start, need_rows));
-            }
-            RETURN_IF_ERROR(_tmp_chunk->capacity_limit_reached());
-        } else {
-            need_rows = std::min(_desired_size, remain_rows);
-            _tmp_chunk = chunk->clone_empty(_desired_size);
-            TRY_CATCH_BAD_ALLOC(_tmp_chunk->append(*chunk, start, need_rows));
-        }
-
-        if (_tmp_chunk->num_rows() >= _desired_size) {
-            _output.emplace_back(std::move(_tmp_chunk));
-        }
-        start += need_rows;
-    }
-    _accumulate_count++;
-    return Status::OK();
-}
-
-bool ChunkAccumulator::empty() const {
-    return _output.empty();
-}
-
-bool ChunkAccumulator::reach_limit() const {
-    return _accumulate_count >= kAccumulateLimit;
-}
-
-ChunkPtr ChunkAccumulator::pull() {
-    if (!_output.empty()) {
-        auto res = std::move(_output.front());
-        _output.pop_front();
-        _accumulate_count = 0;
-        return res;
-    }
-    return nullptr;
-}
-
-void ChunkAccumulator::finalize() {
-    if (_tmp_chunk) {
-        _output.emplace_back(std::move(_tmp_chunk));
-    }
-    _accumulate_count = 0;
 }
 
 bool ChunkPipelineAccumulator::_check_json_schema_equallity(const Chunk* one, const Chunk* two) {
