+++ be/test/runtime/chunk_accumulator_test.cpp
@@ -1,0 +1,80 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "runtime/chunk_accumulator.h"
+
+#include <memory>
+
+#include "column/chunk.h"
+#include "column/fixed_length_column.h"
+#include "gtest/gtest.h"
+
+namespace starrocks {
+
+namespace {
+
+ChunkPtr make_chunk(size_t num_rows) {
+    auto chunk = std::make_shared<Chunk>();
+    auto column = Int32Column::create();
+    column->append_default(num_rows);
+    chunk->append_column(std::move(column), 0);
+    return chunk;
+}
+
+} // namespace
+
+TEST(ChunkAccumulatorTest, Accumulator) {
+    constexpr size_t kDesiredSize = 4096;
+    ChunkAccumulator accumulator(kDesiredSize);
+    size_t input_rows = 0;
+    size_t output_rows = 0;
+
+    // push small chunks
+    for (int i = 0; i < 10; i++) {
+        auto chunk = make_chunk(1025);
+        input_rows += 1025;
+
+        static_cast<void>(accumulator.push(std::move(chunk)));
+        if (ChunkPtr output = accumulator.pull()) {
+            output_rows += output->num_rows();
+            EXPECT_EQ(kDesiredSize, output->num_rows());
+        }
+    }
+
+    // push large chunks
+    for (int i = 0; i < 10; i++) {
+        auto chunk = make_chunk(8888);
+        input_rows += 8888;
+        static_cast<void>(accumulator.push(std::move(chunk)));
+    }
+
+    accumulator.finalize();
+    while (ChunkPtr output = accumulator.pull()) {
+        EXPECT_LE(output->num_rows(), kDesiredSize);
+        output_rows += output->num_rows();
+    }
+    EXPECT_EQ(input_rows, output_rows);
+
+    // push empty chunks
+    for (int i = 0; i < ChunkAccumulator::kAccumulateLimit; i++) {
+        auto chunk = make_chunk(0);
+        static_cast<void>(accumulator.push(std::move(chunk)));
+    }
+    EXPECT_TRUE(accumulator.reach_limit());
+    auto output = accumulator.pull();
+    EXPECT_EQ(nullptr, output);
+    EXPECT_TRUE(accumulator.reach_limit());
+}
+
+} // namespace starrocks
